refactor: move os-specific code to their own files
