+++ src/main.rs
@@ -15,64 +15,9 @@ mod shared {
     }
 }
 
-#[cfg(target_os = "macos")]
-mod platform {
-    use icrate::AppKit::NSScreen;
-    use icrate::Foundation::CGRect;
-    use crate::shared::Screen;
-
-    pub fn browser_path() -> String {
-        String::from("/Applications/Google Chrome.app/Contents/MacOS/Google Chrome")
-    }
-    
-    pub fn displays() -> Vec<Screen> {
-        let mut connected : Vec<Screen> = Vec::new();
-        unsafe {
-            let screens = NSScreen::screens();
-            for (id, screen) in screens.iter().enumerate() {
-                let CGRect { origin, size, .. } = screen.frame();
-                // Flip coordinate space because Apple is weird
-                // https://developer.apple.com/documentation/coregraphics/cgrect
-                let mut origin_y = origin.y as i32;
-                let height = size.height as u32;
-                if connected.len() > 0 {
-                    origin_y = -(height as i32) - (origin_y - connected[0].y);
-                }
-                connected.push(Screen {
-                    id: id as u8,
-                    width: size.width as u32,
-                    height,
-                    x: origin.x as i32,
-                    y: origin_y,
-                })
-            }
-        }
-        connected
-    }
-}
-
-#[cfg(target_os = "linux")]
-mod platform {
-    use std::process::Command;
-
-    const POSSIBLE_BROWSER_BINARIES: [&str; 3] = [
-        "chromium-browser",    // ubuntu/debian
-        "chromium",            // arch
-        "google-chrome-stable" // arch
-    ];
-    pub fn browser_path() -> String {
-        for binary in POSSIBLE_BROWSER_BINARIES {
-            let which = Command::new("which")
-                .arg(binary)
-                .output()
-                .expect("unable to get browser location!");
-            let location = String::from_utf8_lossy(&which.stdout);
-            //dbg!(location);
-            return location.trim_end_matches("\n").to_string();
-        }
-        panic!("No suitable browser found to launch!");
-    }
-}
+#[cfg_attr(target_os = "macos", path = "platforms/macos.rs")]
+#[cfg_attr(target_os = "linux", path = "platforms/linux.rs")]
+mod platform;
 
 use crate::shared::Screen;
